// src/types/models/TransactionPayerDetails.ts
/**
 * Yapily API
 * The Yapily API enables connections between your application and users\' banks. For more information check out our [documentation](https://docs.yapily.com/).<br><br>In particular, make sure to view our [Getting Started](https://docs.yapily.com/pages/home/getting-started/) steps if this is your first time here.<br><br>While testing the API, our list of [sandbox credentials](https://docs.yapily.com/pages/key-concepts/sandbox-credentials/) maybe useful.
 *
 * OpenAPI spec version: 2.30.1
 * Contact: support@yapily.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { TransactionPayeeDetailsAccountIdentificationsInner } from '../models/TransactionPayeeDetailsAccountIdentificationsInner';
import { HttpFile } from '../http/http';

/**
* Details of the benefactor [person or business].
*/
export class TransactionPayerDetails {
    /**
    * The account holder name of the Payer.
    */
    'name'?: string;
    /**
    * The account identifications that identify the Payer\'s bank account.
    */
    'accountIdentifications'?: Array<TransactionPayeeDetailsAccountIdentificationsInner>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "accountIdentifications",
            "baseName": "accountIdentifications",
            "type": "Array<TransactionPayeeDetailsAccountIdentificationsInner>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return TransactionPayerDetails.attributeTypeMap;
    }

    public constructor() {
    }
}