// src/types/models/Enrichment.ts
/**
 * Yapily API
 * The Yapily API enables connections between your application and users\' banks. For more information check out our [documentation](https://docs.yapily.com/).<br><br>In particular, make sure to view our [Getting Started](https://docs.yapily.com/pages/home/getting-started/) steps if this is your first time here.<br><br>While testing the API, our list of [sandbox credentials](https://docs.yapily.com/pages/key-concepts/sandbox-credentials/) maybe useful.
 *
 * OpenAPI spec version: 2.30.1
 * Contact: support@yapily.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Categorisation } from '../models/Categorisation';
import { EnrichmentMerchant } from '../models/EnrichmentMerchant';
import { TransactionHash } from '../models/TransactionHash';
import { HttpFile } from '../http/http';

/**
* Enriched data that has been derived by Yapily using it\'s data processing and machine learning techniques.
*/
export class Enrichment {
    'categorisation'?: Categorisation;
    'transactionHash'?: TransactionHash;
    /**
    * Cleaned version of the `Transaction Description` that removes miscellaneous, generic and unhelpful text.
    */
    'cleansedDescription'?: string;
    'merchant'?: EnrichmentMerchant;
    /**
    * The location of where the transaction took place.
    */
    'location'?: string;
    /**
    * A payment provider that manages (credit/debit) transactions between the `Institution` and the merchant.
    */
    'paymentProcessor'?: string;
    /**
    * The likely date and time on which the transaction took place. This is distinct from `Booking Date Time` which usually refers to the post-clearing value.
    */
    'correctedDate'?: Date;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "categorisation",
            "baseName": "categorisation",
            "type": "Categorisation",
            "format": ""
        },
        {
            "name": "transactionHash",
            "baseName": "transactionHash",
            "type": "TransactionHash",
            "format": ""
        },
        {
            "name": "cleansedDescription",
            "baseName": "cleansedDescription",
            "type": "string",
            "format": ""
        },
        {
            "name": "merchant",
            "baseName": "merchant",
            "type": "EnrichmentMerchant",
            "format": ""
        },
        {
            "name": "location",
            "baseName": "location",
            "type": "string",
            "format": ""
        },
        {
            "name": "paymentProcessor",
            "baseName": "paymentProcessor",
            "type": "string",
            "format": ""
        },
        {
            "name": "correctedDate",
            "baseName": "correctedDate",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return Enrichment.attributeTypeMap;
    }

    public constructor() {
    }
}