// src/types/models/ProprietaryBankTransactionCode.ts
/**
 * Yapily API
 * The Yapily API enables connections between your application and users\' banks. For more information check out our [documentation](https://docs.yapily.com/).<br><br>In particular, make sure to view our [Getting Started](https://docs.yapily.com/pages/home/getting-started/) steps if this is your first time here.<br><br>While testing the API, our list of [sandbox credentials](https://docs.yapily.com/pages/key-concepts/sandbox-credentials/) maybe useful.
 *
 * OpenAPI spec version: 2.30.1
 * Contact: support@yapily.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* Transaction code that is proprietary to the `Institution`.
*/
export class ProprietaryBankTransactionCode {
    /**
    * __Mandatory__. Properietary code used to identify the underlying transaction.
    */
    'code'?: string;
    /**
    * __Mandatory__. Issuer of the properitary code.
    */
    'issuer'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "code",
            "baseName": "code",
            "type": "string",
            "format": ""
        },
        {
            "name": "issuer",
            "baseName": "issuer",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ProprietaryBankTransactionCode.attributeTypeMap;
    }

    public constructor() {
    }
}